1.0版本基本完成

+++ src/main/java/com/jhzz/myblog/controller/CommentController.java
@@ -1,12 +1,16 @@
 package com.jhzz.myblog.controller;
 
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+import com.jhzz.myblog.aop.LogAnnotation;
+import com.jhzz.myblog.common.AppHttpCodeEnum;
 import com.jhzz.myblog.common.ResponseResult;
 import com.jhzz.myblog.domain.Comment;
 import com.jhzz.myblog.domain.param.CommentParam;
 import com.jhzz.myblog.service.CommentService;
+import com.jhzz.myblog.service.LoginService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -24,10 +28,12 @@ import java.util.List;
 @Api
 @RestController
 @RequestMapping("comment")
+@Slf4j
 public class CommentController {
     @Autowired
     private CommentService commentService;
-
+    @Autowired
+    private LoginService loginService;
     @GetMapping("get")
     public ResponseResult getCommentList(@RequestParam("id")Long id,@RequestParam("page")Long page){
         return commentService.getCommentList(id, page);
@@ -35,13 +41,22 @@ public class CommentController {
     }
     @ApiOperation("评论文章")
     @PostMapping("/add")
-    public ResponseResult comment(@RequestBody CommentParam commentParam) {
+    @LogAnnotation(module = "评论模块",operation = "新增评论")
+    public ResponseResult comment(@RequestBody CommentParam commentParam,@RequestHeader("Authorization") String token) {
+        ResponseResult result = loginService.checkToken(token);
+        if (!result.getCode().equals(AppHttpCodeEnum.SUCCESS)){
+            return ResponseResult.errorResult(AppHttpCodeEnum.NO_OPERATOR_AUTH,AppHttpCodeEnum.NO_OPERATOR_AUTH.getMsg());
+        }
         return commentService.comment(commentParam);
     }
 
     @ApiOperation("删除评论")
     @PostMapping("/delete")
-    public ResponseResult deleteComment(@RequestBody CommentParam commentParam) {
+    public ResponseResult deleteComment(@RequestBody CommentParam commentParam,@RequestHeader("Authorization") String token) {
+        ResponseResult result = loginService.checkToken(token);
+        if (!result.getCode().equals(AppHttpCodeEnum.SUCCESS)){
+            return ResponseResult.errorResult(AppHttpCodeEnum.NO_OPERATOR_AUTH,AppHttpCodeEnum.NO_OPERATOR_AUTH.getMsg());
+        }
         return commentService.deleteComment(commentParam);
     }
