+++ src/main/java/com/jhzz/myblog/controller/OSSController.java
@@ -1,10 +1,14 @@
 package com.jhzz.myblog.controller;
 
 import com.jhzz.myblog.aop.LogAnnotation;
+import com.jhzz.myblog.common.AppHttpCodeEnum;
 import com.jhzz.myblog.common.ResponseResult;
+import com.jhzz.myblog.service.LoginService;
 import com.jhzz.myblog.service.OssService;
 import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -20,29 +24,41 @@ import javax.annotation.Resource;
  */
 @RestController
 @Slf4j
-@Api
 @RequestMapping("oss")
 public class OSSController {
     @Resource
     private OssService ossService;
+    @Autowired
+    private LoginService loginService;
 
     @PostMapping("/upload")
-    public ResponseResult uploadFile(@RequestParam("file") MultipartFile file) {
-        log.info("file:{}",file.getOriginalFilename());
+    @LogAnnotation(module = "OSS", operation = "上传图片")
+    public ResponseResult uploadFile(@RequestParam("file") MultipartFile file, @RequestHeader("Authorization") String token) {
+        ResponseResult result = loginService.checkToken(token);
+        if (!result.getCode().equals(200)) {
+            return ResponseResult.errorResult(AppHttpCodeEnum.NO_OPERATOR_AUTH, AppHttpCodeEnum.NO_OPERATOR_AUTH.getMsg());
+        }
+        log.info("file:{}", file.getOriginalFilename());
         String url = ossService.uploadFile(file);
         return ResponseResult.okResult(url);
     }
+
     @PostMapping("/uploadlist")
-    public ResponseResult uploadFileList(@RequestParam("file[]") MultipartFile file) {
-        log.info("file:{}",file.getOriginalFilename());
+    public ResponseResult uploadFileList(@RequestParam("file") MultipartFile file) {
+        log.info("file:{}", file.getOriginalFilename());
         String url = ossService.uploadFile(file);
         return ResponseResult.okResult(url);
     }
+
     @PostMapping("/delete")
-    public ResponseResult deleteFile(@RequestParam("filename") String filename) {
-        log.info("filename:{}",filename);
+    public ResponseResult deleteFile(@RequestParam("filename") String filename, @RequestHeader("Authorization") String token) {
+        ResponseResult result = loginService.checkToken(token);
+        if (!result.getCode().equals(200)) {
+            return ResponseResult.errorResult(AppHttpCodeEnum.NO_OPERATOR_AUTH, AppHttpCodeEnum.NO_OPERATOR_AUTH.getMsg());
+        }
+        log.info("filename:{}", filename);
         ossService.deleteFile(filename);
-        return ResponseResult.okResult(200,"未保存，删除图片");
+        return ResponseResult.okResult(200, "未保存，删除图片");
     }
 
 }
